[Iluvatar GPU] Adapt VL model

+++ ernie/moe/moe_all_gather_layer.py
@@ -367,7 +367,9 @@ class AlltoAllSmart(paddle.autograd.PyLayer):
                     .reshape([-1])
                 )
                 distributed_input_to_alltoall_out = paddle.maximum(
-                    recv_mask_alltoall_out.cumsum() - 1,
+                    (recv_mask_alltoall_out.cumsum() - 1).astype(
+                        recv_mask_alltoall_out.dtype
+                    ),
                     paddle.zeros([1], dtype=recv_mask_alltoall_out.dtype),
                 )
                 distributed_input_to_alltoall_out = (
